Refactor RBAC panel hooks
